Refactor event predicate logics
