+++ ThaiTechEventsCalendar/Networks/CalendarAPI.swift
@@ -11,6 +11,7 @@ import RealmSwift
 import SwiftyJSON
 import Alamofire_SwiftyJSON
 import Alamofire
+import CVCalendar
 
 struct TTConstant {
     static let baseAPIURL = "https://thaiprogrammer-tech-events-calendar.spacet.me"
@@ -35,5 +36,39 @@ struct CalendarAPI {
                 json.arrayValue.forEach({ realm.add(Event($0), update: true) })
             }
         }
+    }
+
+    func events(on date: Date) -> Results<Event>? {
+        guard let realm = try? Realm() else {
+            return nil
+        }
+        let tmr = date.addingTimeInterval(24 * 60 * 60)
+        let thatDatePredicate = NSPredicate(format: "start >= %@ AND end <= %@", date as NSDate, tmr as NSDate)
+        return realm
+            .objects(Event.self)
+            .filter(thatDatePredicate)
+            .sorted(byKeyPath: "start", ascending: true)
+    }
+
+    func upcomingEvents() -> Results<Event>? {
+        guard let realm = try? Realm() else {
+            return nil
+        }
+        let upcomingPredicate = NSPredicate(format: "start >= %@", Date() as NSDate)
+        return realm
+            .objects(Event.self)
+            .filter(upcomingPredicate)
+            .sorted(byKeyPath: "start", ascending: true)
+    }
+
+    func pastEvents() -> Results<Event>? {
+        guard let realm = try? Realm() else {
+            return nil
+        }
+        let pastPredicate = NSPredicate(format: "start < %@", Date() as NSDate)
+        return realm
+            .objects(Event.self)
+            .filter(pastPredicate)
+            .sorted(byKeyPath: "start", ascending: true)
     }
 }
